+++ plugins/woodpecker-cmip6-decadal-plugin/src/woodpecker_cmip6_decadal_plugin/__init__.py
@@ -15,4 +15,3 @@ from .cmip6d_0012 import DecadalLeadtimeMetadataNormalizationFix  # noqa: F401
 from .cmip6d_0013 import DecadalModelGlobalAttributesFix  # noqa: F401
 from .cmip6d_0014 import DecadalReftimeCoordinateFix  # noqa: F401
 from .cmip6d_0015 import DecadalLeadtimeCoordinateFix  # noqa: F401
-from .cmip6d_0999 import Cmip6DecadalFullFixSuiteFix  # noqa: F401